project-07-cross-cloud-compliance/pyproject.toml
[project]
name = "cross-cloud-compliance"
version = "0.1.0"
description = "A2A-federated compliance auditor with GDPR/SOX/HIPAA agents across AWS/GCP/Azure"
requires-python = ">=3.11"
dependencies = [
    "a2a-sdk>=0.3.0",
    "anthropic>=0.40.0",
    "langgraph>=0.4.0",
    "langchain-anthropic>=0.3.0",
    "boto3>=1.34.0",
    "google-generativeai>=0.8.0",
    "openai>=1.50.0",
    "httpx>=0.27.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "python-multipart>=0.0.12",
    "pypdf>=4.0.0",
    "pydantic-settings>=2.4.0",
    "python-dotenv>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
